+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - add Event.getRow() and Row.name
 - expose domain changes, bound changes, branching decisions, added constraints for nodes
 - define Python object identity based on underlying SCIP object pointers, so that e.g. rows and columns can be added to sets, and testing for equality is consistent over time.
+- add Row.isRemovable and Row.getOrigintype
 
 ## 2.2.3 - 2019-12-10
 ### Added